fix: apply rate limiting globally and harden Dockerfiles
