+++ Dockerfile
@@ -36,8 +36,10 @@ RUN addgroup -g 1001 -S archivum && \
 WORKDIR /app
 
 # Install production dependencies only
-COPY backend/package.json backend/package-lock.json* backend/.npmrc* ./backend/
-RUN cd backend && npm ci --omit=dev --ignore-scripts && npm cache clean --force
+WORKDIR /app/backend
+COPY backend/package.json backend/package-lock.json* backend/.npmrc* ./
+RUN npm ci --omit=dev --ignore-scripts && npm cache clean --force
+WORKDIR /app
 
 # Copy built backend
 COPY --from=backend-build /app/backend/dist ./backend/dist
